+++ src/modules/server/server.controller.ts
@@ -28,4 +28,10 @@ export class ServerController {
     return this.serverService.resetHomebridgeAccessory();
   }
 
+  @UseGuards(AdminGuard)
+  @Put('/remove-orphans')
+  removeOrphans() {
+    return this.serverService.removeOrphans();
+  }
+
 }